Fix: Refactored code. Added IPFS upload
